Hooking working again. hook.cpp clashes with hook.c. Need separate directories for .obj files..
